Extract stateless condition type checkers to validation.py
