+++ nucypher/policy/conditions/evm.py
@@ -4,20 +4,16 @@ from typing import (
     Iterator,
     List,
     Optional,
-    Sequence,
     Set,
     Tuple,
     Type,
     Union,
-    cast,
 )
 
 from eth_typing import ChecksumAddress
 from eth_utils import to_checksum_address
-from hexbytes import HexBytes
 from marshmallow import ValidationError, fields, post_load, validate, validates_schema
 from web3 import HTTPProvider, Web3
-from web3.auto import w3
 from web3.contract.contract import ContractFunction
 from web3.middleware import geth_poa_middleware
 from web3.providers import BaseProvider
@@ -34,6 +30,12 @@ from nucypher.policy.conditions.exceptions import (
 )
 from nucypher.policy.conditions.lingo import ConditionType, ReturnValueTest
 from nucypher.policy.conditions.utils import CamelCaseSchema, camel_case_to_snake
+from nucypher.policy.conditions.validation import (
+    _align_comparator_value_with_abi,
+    _get_abi_types,
+    _validate_multiple_output_types,
+    _validate_single_output_type,
+)
 
 # TODO: Move this to a more appropriate location,
 #  but be sure to change the mocks in tests too.
@@ -369,41 +371,8 @@ class ContractCondition(RPCCondition):
     def _validate_method(self, method):
         return method
 
-    def _validate_single_output_type(
-        self,
-        expected_type: str,
-        comparator_value: Any,
-        comparator_index: Optional[int],
-        failure_message: str,
-    ) -> None:
-        if comparator_index is not None and self._is_tuple_type(expected_type):
-            type_entries = self._get_tuple_type_entries(expected_type)
-            expected_type = type_entries[comparator_index]
-        self._validate_value_type(expected_type, comparator_value, failure_message)
-
-    def _validate_multiple_output_types(
-        self,
-        output_abi_types: List[str],
-        comparator_value: Any,
-        comparator_index: Optional[int],
-        failure_message: str,
-    ) -> None:
-        if comparator_index is not None:
-            expected_type = output_abi_types[comparator_index]
-            self._validate_value_type(expected_type, comparator_value, failure_message)
-            return
-
-        if not isinstance(comparator_value, Sequence):
-            raise InvalidCondition(failure_message)
-
-        if len(output_abi_types) != len(comparator_value):
-            raise InvalidCondition(failure_message)
-
-        for output_abi_type, component_value in zip(output_abi_types, comparator_value):
-            self._validate_value_type(output_abi_type, component_value, failure_message)
-
     def _validate_expected_return_type(self) -> None:
-        output_abi_types = self._get_abi_types(self.contract_function.contract_abi[0])
+        output_abi_types = _get_abi_types(self.contract_function.contract_abi[0])
         comparator_value = self.return_value_test.value
         comparator_index = self.return_value_test.index
         index_string = (
@@ -415,39 +384,15 @@ class ContractCondition(RPCCondition):
         )
 
         if len(output_abi_types) == 1:
-            self._validate_single_output_type(
+            _validate_single_output_type(
                 output_abi_types[0], comparator_value, comparator_index, failure_message
             )
         elif len(output_abi_types) > 1:
-            self._validate_multiple_output_types(
+            _validate_multiple_output_types(
                 output_abi_types, comparator_value, comparator_index, failure_message
             )
         else:
             raise InvalidCondition("No output types available for ABI function.")
-
-    def _validate_value_type(
-        self, expected_type: str, comparator_value: Any, failure_message: str
-    ) -> None:
-        if is_context_variable(comparator_value):
-            # context variable types cannot be known until execution time.
-            return
-
-        comparator_value = self._align_comparator_value(
-            comparator_value, expected_type, failure_message
-        )
-        if not w3.is_encodable(expected_type, comparator_value):
-            raise InvalidCondition(failure_message)
-
-    @staticmethod
-    def _align_comparator_value(
-        comparator_value: Any, expected_type: str, failure_message: str
-    ) -> Any:
-        if expected_type.startswith("bytes"):
-            try:
-                comparator_value = bytes(HexBytes(comparator_value))
-            except Exception:
-                raise InvalidCondition(failure_message)
-        return comparator_value
 
     def __repr__(self) -> str:
         r = (
@@ -491,90 +436,7 @@ class ContractCondition(RPCCondition):
     def _align_comparator_value_with_abi(
         self, return_value_test: ReturnValueTest
     ) -> ReturnValueTest:
-        output_abi_types = self._get_abi_types(self.contract_function.contract_abi[0])
-        comparator = return_value_test.comparator
-        comparator_value = return_value_test.value
-        comparator_index = return_value_test.index
-        if isinstance(comparator_value, tuple):
-            # must be list
-            comparator_value = list(comparator_value)
-
-        if len(output_abi_types) == 1:
-            expected_type = output_abi_types[0]
-            if comparator_index is not None and self._is_tuple_type(expected_type):
-                type_entries = self._get_tuple_type_entries(expected_type)
-                expected_type = type_entries[comparator_index]
-            comparator_value = self._align_comparator_value(
-                comparator_value, expected_type, failure_message="Unencodable type"
-            )
-            return ReturnValueTest(
-                comparator=comparator,
-                value=comparator_value,
-                index=comparator_index,
-            )
-        elif len(output_abi_types) > 1:
-            if comparator_index is not None:
-                # only index entry we care about
-                expected_type = output_abi_types[comparator_index]
-                comparator_value = self._align_comparator_value(
-                    comparator_value,
-                    expected_type,
-                    failure_message="Unencodable type",
-                )
-                return ReturnValueTest(
-                    comparator=comparator,
-                    value=comparator_value,
-                    index=comparator_index,
-                )
-
-            values = list()
-            for output_abi_type, component_value in zip(
-                output_abi_types, comparator_value
-            ):
-                comparator_value = self._align_comparator_value(
-                    comparator_value,
-                    output_abi_type,
-                    failure_message=f"Unencodable type ({comparator_value} as {output_abi_type})",
-                )
-                values.append(component_value)
-            return ReturnValueTest(
-                comparator=comparator, value=values, index=comparator_index
-            )
-        else:
-            # should never happen (validated on object initialization)
-            raise RuntimeError("No output types available for ABI function.")
-
-    @classmethod
-    def _get_abi_types(cls, abi: ABIFunction) -> List[str]:
-        if abi["type"] == "fallback":
-            return []
-        else:
-            return [
-                cls._collapse_if_tuple(cast(Dict[str, Any], arg))
-                for arg in abi["outputs"]
-            ]
-
-    @classmethod
-    def _collapse_if_tuple(cls, abi: Dict[str, Any]) -> str:
-        abi_type = abi["type"]
-        if not abi_type.startswith("tuple"):
-            return abi_type
-
-        delimited = ",".join(cls._collapse_if_tuple(c) for c in abi["components"])
-        collapsed = f"({delimited})"
-        return collapsed
-
-    @classmethod
-    def _is_tuple_type(cls, abi_type: str):
-        return abi_type.startswith("(") and abi_type.endswith(")")
-
-    @classmethod
-    def _get_tuple_type_entries(cls, tuple_type: str) -> List[str]:
-        if not cls._is_tuple_type(tuple_type):
-            raise ValueError(
-                f"Invalid type provided '{tuple_type}; not a tuple type definition"
-            )
-
-        result = tuple_type.replace("(", "").replace(")", "")
-        result = result.split(",")
-        return result
+        return _align_comparator_value_with_abi(
+            abi=self.contract_function.contract_abi[0],
+            return_value_test=return_value_test,
+        )
